move EventConfig into event stream module
