+++ platform-sdk/consensus-event-stream/build.gradle.kts
@@ -8,19 +8,22 @@ plugins {
 
 description = "Consensus event-stream file writing"
 
+mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
+
 testModuleInfo {
     requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.event.stream.test.fixtures")
+    requires("org.hiero.consensus.utility")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
     requires("org.mockito")
 }
 
 timingSensitiveModuleInfo {
+    requires("org.hiero.base.crypto")
     requires("org.hiero.consensus.event.stream")
     requires("org.hiero.consensus.event.stream.test.fixtures")
-    requires("org.hiero.base.crypto")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
 }
