+++ platform-sdk/consensus-event-stream/src/main/java/module-info.java
@@ -3,10 +3,15 @@ import com.swirlds.config.api.ConfigurationExtension;
 import org.hiero.consensus.event.stream.config.EventStreamConfigurationExtension;
 
 module org.hiero.consensus.event.stream {
-    exports org.hiero.consensus.event.stream;
     exports org.hiero.consensus.event.stream.config;
+    exports org.hiero.consensus.event.stream;
     exports org.hiero.consensus.event.stream.internal to
             org.hiero.consensus.event.stream.test.fixtures;
+
+    opens org.hiero.consensus.event.stream to
+            com.fasterxml.jackson.databind;
+    opens org.hiero.consensus.event.stream.internal to
+            com.fasterxml.jackson.databind;
 
     requires transitive com.swirlds.base;
     requires transitive com.swirlds.component.framework;
@@ -20,15 +25,9 @@ module org.hiero.consensus.event.stream {
     requires com.swirlds.logging;
     requires org.hiero.base.concurrent;
     requires org.hiero.consensus.metrics;
-    requires org.hiero.consensus.utility;
     requires org.apache.logging.log4j;
     requires static transitive com.github.spotbugs.annotations;
 
     provides ConfigurationExtension with
             EventStreamConfigurationExtension;
-
-    opens org.hiero.consensus.event.stream to
-            com.fasterxml.jackson.databind;
-    opens org.hiero.consensus.event.stream.internal to
-            com.fasterxml.jackson.databind;
 }
