+++ platform-sdk/consensus-event-stream/src/main/java/org/hiero/consensus/event/stream/DefaultConsensusEventStream.java
@@ -24,7 +24,7 @@ import org.hiero.base.crypto.DigestType;
 import org.hiero.base.crypto.Hash;
 import org.hiero.base.crypto.Signer;
 import org.hiero.consensus.concurrent.throttle.RateLimitedLogger;
-import org.hiero.consensus.config.EventConfig;
+import org.hiero.consensus.event.stream.config.EventConfig;
 import org.hiero.consensus.event.stream.internal.HashCalculatorForStream;
 import org.hiero.consensus.event.stream.internal.QueueThreadObjectStream;
 import org.hiero.consensus.event.stream.internal.QueueThreadObjectStreamConfiguration;
