+++ platform-sdk/consensus-event-stream/src/main/java/org/hiero/consensus/event/stream/config/EventConfig.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.config;
+package org.hiero.consensus.event.stream.config;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
@@ -17,5 +17,8 @@ import com.swirlds.config.api.ConfigProperty;
 public record EventConfig(
         @ConfigProperty(defaultValue = "5000") int eventStreamQueueCapacity,
         @ConfigProperty(defaultValue = "5") long eventsLogPeriod,
-        @ConfigProperty(defaultValue = "/opt/hgcapp/eventsStreams") String eventsLogDir,
+
+        @ConfigProperty(defaultValue = "/opt/hgcapp/eventsStreams")
+        String eventsLogDir,
+
         @ConfigProperty(defaultValue = "true") boolean enableEventStreaming) {}
