+++ platform-sdk/consensus-event-stream/src/main/java/org/hiero/consensus/event/stream/config/EventStreamConfigurationExtension.java
@@ -16,7 +16,7 @@ public class EventStreamConfigurationExtension implements ConfigurationExtension
      */
     @NonNull
     public Set<Class<? extends Record>> getConfigDataTypes() {
-        return Set.of(EventStreamWiringConfig.class);
+        return Set.of(EventConfig.class, EventStreamWiringConfig.class);
     }
 
     /**
