+++ platform-sdk/consensus-event-stream/src/testFixtures/java/module-info.java
@@ -16,7 +16,6 @@ open module org.hiero.consensus.event.stream.test.fixtures {
     requires org.hiero.base.utility.test.fixtures;
     requires org.hiero.consensus.concurrent;
     requires org.hiero.consensus.model.test.fixtures;
-    requires org.hiero.consensus.utility;
     requires org.apache.logging.log4j;
     requires org.mockito;
 }
