+++ platform-sdk/consensus-event-stream/src/testFixtures/java/org/hiero/consensus/event/stream/test/fixtures/EventStreamTestUtils.java
@@ -25,17 +25,18 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.hiero.base.crypto.DigestType;
 import org.hiero.base.crypto.Hash;
 import org.hiero.base.crypto.Signer;
-import org.hiero.consensus.config.EventConfig_;
 import org.hiero.consensus.event.stream.DefaultConsensusEventStream;
 import org.hiero.consensus.event.stream.EventStreamType;
 import org.hiero.consensus.event.stream.LinkedObjectStream;
 import org.hiero.consensus.event.stream.RunningHashCalculatorForStream;
+import org.hiero.consensus.event.stream.config.EventConfig_;
 import org.hiero.consensus.event.stream.internal.TimestampStreamFileWriter;
 import org.hiero.consensus.metrics.noop.NoOpMetrics;
 import org.hiero.consensus.model.event.CesEvent;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.hashgraph.ConsensusRound;
 import org.hiero.consensus.model.node.NodeId;
+// import org.hiero.consensus.event.stream.config.EventConfig_;
 import org.hiero.consensus.model.test.fixtures.event.TestingEventBuilder;
 import org.mockito.Mockito;
