+++ platform-sdk/consensus-otter-tests/build.gradle.kts
@@ -40,6 +40,7 @@ testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
     requires("com.swirlds.component.framework")
     requires("com.swirlds.metrics.api")
+    requires("org.hiero.consensus.event.stream")
     requires("org.hiero.consensus.metrics")
     requires("org.hiero.consensus.roster")
     requires("org.hiero.consensus.roster.test.fixtures")
