+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/module-info.java
@@ -80,6 +80,7 @@ module org.hiero.otter.fixtures {
     requires com.swirlds.virtualmap;
     requires org.hiero.base.concurrent;
     requires org.hiero.consensus.concurrent;
+    requires org.hiero.consensus.event.stream;
     requires org.hiero.consensus.hashgraph.impl.test.fixtures;
     requires org.hiero.consensus.hashgraph;
     requires org.hiero.consensus.iss.detection;
