+++ platform-sdk/consensus-sloth/src/testFixtures/java/module-info.java
@@ -62,6 +62,7 @@ module org.hiero.sloth.fixtures {
     requires com.swirlds.platform.core.test.fixtures;
     requires org.hiero.base.utility;
     requires org.hiero.consensus.concurrent;
+    requires org.hiero.consensus.event.stream;
     requires org.hiero.consensus.hashgraph;
     requires org.hiero.consensus.metrics;
     requires org.hiero.consensus.platformstate;
