+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/config/UtilityConfigurationExtension.java
@@ -15,11 +15,6 @@ public class UtilityConfigurationExtension implements ConfigurationExtension {
      */
     @NonNull
     public Set<Class<? extends Record>> getConfigDataTypes() {
-        return Set.of(
-                BasicConfig.class,
-                EventConfig.class,
-                PathsConfig.class,
-                PlatformStatusConfig.class,
-                RecycleBinConfig.class);
+        return Set.of(BasicConfig.class, PathsConfig.class, PlatformStatusConfig.class, RecycleBinConfig.class);
     }
 }
