+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/CleanCommand.java
@@ -11,8 +11,8 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Objects;
 import org.hiero.base.file.FileUtils;
-import org.hiero.consensus.config.EventConfig;
 import org.hiero.consensus.config.PathsConfig;
+import org.hiero.consensus.event.stream.config.EventConfig;
 import picocli.CommandLine;
 
 @CommandLine.Command(
